Agrego cambios del esquema de Base de Datos.

+++ ServiciosParlamentarios/apirest/models/expedientes/expediente_origina_despacho.py
@@ -6,7 +6,7 @@ class ExpedienteOriginaDespacho(models.Model):
     id = models.AutoField(primary_key=True,db_column='expediente_origina_despacho_id')
     fk_expediente = models.ForeignKey('Proyecto', db_column='fk_expediente')
     fk_despacho = models.ForeignKey('Despacho', db_column='fk_despacho')
-    orden = models.IntegerField(blank=True, null=True)
+    orden = models.SmallIntegerField(blank=True, null=True)
     solo_vista = models.CharField(max_length=1, blank=True,db_column='bsolovista')
     
     class Meta:
